refactor: share model filter hook
